py/sight/requirements.txt
absl-py==1.3.0
astunparse==1.6.3
attrs==22.1.0
avro==1.11.1
cachetools==4.2.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
cryptography==38.0.4
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.6
dm-acme==0.4.0
dm-env==1.5
dm-haiku==0.0.9
dm-launchpad==0.5.0
dm-reverb==0.7.0
dm-sonnet==2.0.1
dm-tree==0.1.8
docstring-parser==0.15
dsub==0.4.7
fastavro==1.7.0
filelock==3.8.0
fire==0.4.0
flatbuffers==1.12
funcsigs==1.0.2
gast==0.4.0
google-api-core==1.33.2
google-api-python-client==1.12.11
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.18.3
google-cloud-bigquery==1.20.0
google-cloud-core==1.7.3
google-cloud-notebooks==1.4.4
google-cloud-pipeline-components==1.0.26
google-cloud-resource-manager==1.6.3
google-cloud-spanner==1.10.0
google-cloud-storage==1.44.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-gcp==0.2.2
grpcio-status==1.48.2
grpcio-tools==1.48.2
gym==0.26.2
gym-notices==0.0.8
h5py==3.7.0
httplib2==0.20.4
idna==3.4
jax==0.4.1
jaxlib==0.4.1
jmp==0.0.2
joblib==1.2.0
jsonschema==3.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kfp==1.8.14
kfp-pipeline-spec==0.1.16
kfp-server-api==1.8.5
kubernetes==18.20.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mock==4.0.3
ndg-httpsclient==0.5.1
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
parameterized==0.8.1
Pillow==9.3.0
platformdirs==2.5.3
portpicker==1.5.2
proto-plus==1.22.1
protobuf==3.20.2
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
PyYAML==5.4.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.8
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sklearn==0.0
strip-hints==0.1.10
tabulate==0.8.9
tenacity==7.0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-probability==0.15.0
termcolor==2.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
trfl==1.2.0
typer==0.6.1
typing_extensions==4.4.0
uritemplate==3.0.1
urllib3==1.26.12
virtualenv==20.16.6
websocket-client==1.4.1
Werkzeug==2.2.2
wrapt==1.14.1
